Separate aggregate specs
- separates aggregate integration specs into their own file.
- Tests written to be more 'rspec' like. Aim is to write tests this way going forward.
